backend-model-development/requirements.txt
scikit-learn
Pillow
pycocotools
torch
torchvision
torchaudio
fvcore
opencv-python
cython
ultralytics
mlflow
playwright
dvc